support doubles on the subs matrix and go cost

+++ features_for_msas/features_for_msas.cpp
@@ -35,7 +35,7 @@ int main(int argc, char* argv[])
     json config = json::parse(f);
     std::vector<EvoModel> models;
     for (const auto& modelConfig : config["models_list"]) {
-        int gapOpenCost = modelConfig["gap_open_cost"].get<int>();
+        double gapOpenCost = modelConfig["gap_open_cost"].get<double>();
         double gapExtendCost = modelConfig["gap_extend_cost"].get<double>();
         std::string substitutionMatrix = modelConfig["matrix_file_name"].get<std::string>();
         EvoModel evoModel(gapOpenCost, gapExtendCost, substitutionMatrix);
